+++ src/main/java/com/example/demo/database/services/vehicle/FleetService.java
@@ -2,8 +2,8 @@ package com.example.demo.database.services.vehicle;
 
 import com.example.demo.database.models.vehicle.Fleet;
 import com.example.demo.database.repositories.vehicle.FleetRepository;
-import com.example.demo.utils.Mapping;
-import com.example.demo.utils.ValidationResponse;
+import com.example.demo.database.models.utils.Mapping;
+import com.example.demo.database.models.utils.ValidationResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -40,6 +40,13 @@ public class FleetService {
 	}
 
 	public void delete(Fleet fleet) {
+		if (fleet == null) {
+			return;
+		}
+
+		if (fleet.getId() == null) {
+			return;
+		}
 		repository.delete(fleet);
 	}
 
@@ -49,9 +56,40 @@ public class FleetService {
 
 	public ValidationResponse validate(Fleet fleet, Mapping mapping) {
 
-		if (fleet.getName() != null) {
-			if (fleet.getName().length() <= 0) {
-				return new ValidationResponse(false, "name cannot be empty");
+		if (fleet == null) {
+			return new ValidationResponse(false, "provided NULL entity");
+		}
+
+		if (mapping.equals(Mapping.POST)) {
+			fleet.setId(null);
+		}
+
+		if (mapping.equals(Mapping.PUT) || mapping.equals(Mapping.PATCH)) {
+			if (fleet.getId() == null) {
+				return new ValidationResponse(false, "ID parameter is required");
+			}
+
+			Fleet fleetFromDatabase = getById(fleet.getId());
+
+			if (fleetFromDatabase == null) {
+				return new ValidationResponse(false, "ID parameter is invalid");
+			}
+		}
+
+		if (mapping.equals(Mapping.POST) || mapping.equals(Mapping.PUT) || mapping.equals(Mapping.PATCH)) {
+			if (fleet.getName() != null) {
+				if (fleet.getName().length() <= 0) {
+					return new ValidationResponse(false, "name cannot be empty");
+				}
+			}
+
+
+			// OTHER VALIDATION
+		}
+
+		if (mapping.equals(Mapping.DELETE)) {
+			if (fleet.getId() == null) {
+				return new ValidationResponse(false, "ID parameter is required");
 			}
 		}
