+++ src/main/java/com/example/demo/database/services/vehicle/TripService.java
@@ -2,6 +2,8 @@ package com.example.demo.database.services.vehicle;
 
 import com.example.demo.database.models.vehicle.Trip;
 import com.example.demo.database.repositories.vehicle.TripRepository;
+import com.example.demo.database.models.utils.Mapping;
+import com.example.demo.database.models.utils.ValidationResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -46,10 +48,73 @@ public class TripService {
 	}
 
 	public void delete(Trip trip) {
+		if (trip == null) {
+			return;
+		}
+
+		if (trip.getId() == null) {
+			return;
+		}
+
 		repository.delete(trip);
 	}
 
 	public void deleteAll() {
 		repository.deleteAll();
 	}
+
+
+    public ValidationResponse validate(Trip trip, Mapping mapping) {
+
+		if (trip == null) {
+			return new ValidationResponse(false, "provided NULL entity");
+		}
+
+		if (mapping.equals(Mapping.POST)) {
+			trip.setId(null);
+		}
+
+		if (mapping.equals(Mapping.PUT) || mapping.equals(Mapping.PATCH)) {
+			if (trip.getId() == null) {
+				return new ValidationResponse(false, "ID parameter is required");
+			}
+
+			Trip tripFromDatabase = getById(trip.getId());
+
+			if (tripFromDatabase == null) {
+				return new ValidationResponse(false, "ID parameter is invalid");
+			}
+		}
+
+		if (mapping.equals(Mapping.POST) || mapping.equals(Mapping.PUT) || mapping.equals(Mapping.PATCH)) {
+			if (trip.getOrigin() != null) {
+				if (trip.getOrigin().length() <= 0) {
+					return new ValidationResponse(false, "origin cannot be empty");
+				}
+			}
+
+			if (trip.getDestination() != null) {
+				if (trip.getDestination().length() <= 0) {
+					return new ValidationResponse(false, "destination cannot be empty");
+				}
+			}
+
+			if (trip.getKilometres_driven() != null) {
+				if (trip.getKilometres_driven() < 0) {
+					return new ValidationResponse(false, "invalid kilometres value");
+				}
+			}
+
+
+			// OTHER BASIC VALIDATION
+		}
+
+		if (mapping.equals(Mapping.DELETE)) {
+			if (trip.getId() == null) {
+				return new ValidationResponse(false, "ID parameter is required");
+			}
+		}
+
+		return new ValidationResponse(true, "validation success");
+    }
 }
